Created a maven project for Metrics and MetricsLite.
Plugin developers can then conveniently include it using the maven shade plugin.
